requirements.txt
pygame-ce
moderngl